Add is_hugetlbfs() to GuestMemoryRegion
Virtio-balloon can release the unused host memory to decrease the memory
usage of the VMM.
Release normal pages and hugetlbfs pages require different operations.
(madvise MADV_DONTNEED and fallocate64 FALLOC_FL_PUNCH_HOLE)

This commit add Add is_hugetlbfs() to GuestMemoryRegion to help
VMM decide if this is a hugetlbfs address or not.

Signed-off-by: Hui Zhu <teawater@antfin.com>